feat: allow sub-second scan interval (min 200 ms)
Scan interval is now a float; minimum lowered from 2 s to 0.2 s, default
1 s, max 600 s. Config + options flow accept decimals. Bump to 1.0.2.

+++ custom_components/dtsu666_meter/coordinator.py
@@ -39,7 +39,7 @@ class Dtsu666Coordinator(DataUpdateCoordinator[dict[str, float]]):
         host: str,
         port: int,
         slave_id: int,
-        scan_interval: int,
+        scan_interval: float,
     ) -> None:
         super().__init__(
             hass,
